Some refactoring
Move the fetching of Bugzilla data to `SignaturesDataFetcher`
Move finding actionable crashes to `SignaturesDataFetcher`
